+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.8] - 2023-10-03
 
 ### Changed
 
 - Headers of .txt and .csv files are checked to match requirements of ORPL.
   - Headers NEED to follow the structutre `xaxis,background,accumulation_0,accumulation_1,...` if header keys are missing or not named correctly, loading will fail.
+  - Headers NEED at least 1 `accumulation_X` key to be valid
+  - Header keys `xaxis` and `background` are optional
+
+### Fixed
+
+- Processing of spectrum from `.txt` and `.csv` files that don't have backgrounds
 
 ## [1.0.7] - 2023-06-21